This is a version of the ethical governor example suitable for exporting a model to Prism.
Note that the end states in the AJPF->Prism model get annotated with the properties from the
previous state even though they have no properties.  This caused some problems in the translation
which were worked around by adding the options to Choice in a particular order.  At some point
the bug needs to be fixed properly.