Fixed a typo in tutorial 2.
Started writing tutorial 8 which is to be on multi-agent systems and communication.

I created a new environment, based on SearchAndRescueEnv.java which includes a human to be assisted and differentiates between a "lifter" robot and a "medic" robot.  I also started writing a simple program for this environment to demonstrate communication.  This program does not appear to be working.