Finally managed to fix the various problems with Logical Consequence and unification.  This time the bugs were mostly in the handling of capability switching, not in the algorithms for reasoning.
I then discovered some bugs in the implementation of feedback control which I
simplified to allow wheels to roll backwards if the robot is facing very much
the wrong way.

It has all worked once.  I need to run the regression tests again and then do
some neatening and commenting.