Initial implementation of replay.
Note: Dysprosium and Tutoria1Tests broken, but want to check against changes in platoon branch before fixing.

Implementation required adding a controller to environments and some changes to setups as a result.
At the moment record is printed out at the end of a run.  Need to get a dump in the event of a verification
fail and, potentially cut and paste print outs.